Revert "RavenDB-17360 move some files around and address PR comments"
This reverts commit 8d5909fdfa99a07a8fde630fafa340e6b82f6bed.
