+++ src/Raven.Server/Documents/Handlers/SubscriptionsHandler.cs
@@ -12,7 +12,6 @@ using Raven.Client.Exceptions.Documents.Subscriptions;
 using Raven.Server.Documents.Includes;
 using Raven.Server.Documents.Replication;
 using Raven.Server.Documents.Subscriptions;
-using Raven.Server.Documents.Subscriptions.SubscriptionProcessor;
 using Raven.Server.Documents.TcpHandlers;
 using Raven.Server.Json;
 using Raven.Server.Routing;
@@ -87,17 +86,18 @@ namespace Raven.Server.Documents.Handlers
                 var sp = Stopwatch.StartNew();
                 var timeLimit = TimeSpan.FromSeconds(GetIntValueQueryString("timeLimit", false) ?? 15);
                 var startEtag = cv.Etag;
+                
+                var fetcher = new DummyDocumentSubscriptionFetcher(Database, state, sub.Collection);
+                fetcher.SetConnectionInfo(new SubscriptionWorkerOptions("dummy"), state, new IPEndPoint(HttpContext.Connection.RemoteIpAddress, HttpContext.Connection.RemotePort));
+                fetcher.AddScript(patch);
 
-                var processor = new TestDocumentsSubscriptionProcessor(Server.ServerStore, Database, state, sub.Collection, new SubscriptionWorkerOptions("dummy"), new IPEndPoint(HttpContext.Connection.RemoteIpAddress, HttpContext.Connection.RemotePort));
-                processor.AddScript(patch);
-
-                using (processor)
+                using (fetcher)
                 await using (var writer = new AsyncBlittableJsonTextWriter(context, ResponseBodyStream()))
                 using (Database.ServerStore.Engine.ContextPool.AllocateOperationContext(out ClusterOperationContext clusterOperationContext))
                 using (clusterOperationContext.OpenReadTransaction())
                 using (context.OpenReadTransaction())
                 {
-                    processor.InitializeForNewBatch(clusterOperationContext, context, includeCmd);
+                    fetcher.Initialize(clusterOperationContext, context, includeCmd);
 
                     writer.WriteStartObject();
                     writer.WritePropertyName("Results");
@@ -108,9 +108,9 @@ namespace Raven.Server.Documents.Handlers
                         var first = true;
                         var lastEtag = startEtag;
                         
-                        processor.SetStartEtag(startEtag);
+                        fetcher.SetStartEtag(startEtag);
 
-                        foreach (var itemDetails in processor.GetBatch())
+                        foreach (var itemDetails in fetcher.Fetch())
                         {
                             if (itemDetails.Doc.Data != null)
                             {
